🚧 style: refactor Product component to use Product type and improve prop structure
