+++ src/containers/ProductsArea/index.tsx
@@ -1,4 +1,9 @@
 import {ProductsArea} from './styled';
+import { Product } from '../../types/Product'; // Adjust the import path as necessary
+
+type Props = {
+  Products: Product[];
+}
 const ProductsAreaComponent = () => {
   return (
     <>
